% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_aphiaid}
\alias{clean_aphiaid}
\title{get aphia ID's}
\usage{
clean_aphiaid(DF, SP_COL, DF_ABBR)
}
\arguments{
\item{DF}{The dataframe to run the function on}

\item{SP_COL}{The column with the species names}

\item{DF_ABBR}{Dataset abbreviation. Must be a character in inverted commas}
}
\value{
Returns the dataframe with the aphia ID's as a new column
}
\description{
This function matches species names in a dataframe to worms id numbers
}
