% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_area}
\alias{clean_area}
\title{Restrict locations to with the eez}
\usage{
clean_area(DF, LON, LAT)
}
\arguments{
\item{DF}{The dataframe to run the function on}

\item{LON}{The column with the longitude values (as decimals). Should be a
character in inverted commas.}

\item{LAT}{The column with the latitude values (as decimals). Should be a
character in inverted commas.}
}
\value{
Returns the dataframe with only the values that fall with the eez
}
\description{
This function test if locations fall within the marine EEZ with a buffer
of 2km, and drops all other values.
}
