% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_accs}
\alias{extract_accs}
\title{Load tables from access}
\usage{
extract_accs(MDBPATH)
}
\arguments{
\item{MDBPATH}{Path to the access database. Remeber to ass the back slash
before the first folder name. If you are working in a project the file path
should only be the part from the project onwards.}
}
\value{
Returns all tables in the access database as r dataframes
with the same name.
}
\description{
This function loads tables in an access database into r as
dataframes with the same name as the access tables.
}
\examples{
# example code - this won't run but is for you to see the correct formatting
 extract_accs("/data/FishMaster.accdb")



}
