% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_dim_date}
\alias{make_dim_date}
\title{make the date table}
\usage{
make_dim_date()
}
\value{
Returns ta table with every date at daily, yearly, and monthly intervals from 1890/01/01
to 2025/01/01.
}
\description{
To make the data table, instead of adding dates from the data we will
create a table with every date over a date range that our data is
expected to cover. This may need to be extended as this work extends passed
2025, but extra rows can be added easily enough.
It is easier to make the table in R so that we can use the lubridate functions
}
\examples{
## example code
# date_tbl <- make_dim_date

}
