% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_dim_taxon}
\alias{make_dim_taxon}
\title{make taxon table}
\usage{
make_dim_taxon(DF)
}
\arguments{
\item{DF}{The dataframe to run the function on}
}
\value{
Returns a table of taxon information for each taxon records in the dataset
}
\description{
To make the taxon table use classification function to retrieve the taxonomic information for each aphia
ID from WORMS.
Then use rbind to take the information from a list format to a database.
Then remove the id column as it is not needed.
The data is in long format, to get it into wide format first make sure that each aphiaID (query), and
rank (kingdom/family ect) only has one name (eg agryrosomus) associated with it. Basically remove any
duplicates.
Then convert it to wide format and rename query as aphiaid.
Get a dataframe of the unique aphiaid's with their associated taxon names from the original dataset, removing
any species that don't have an aphiaid.
Join the WORMS taxon table to the dataframe.
}
\examples{
## example code
# fm_taxon <- make_dim_taxon(taxon_new)

}
