% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_dis_spp}
\alias{make_dis_spp}
\title{get a list of distinct species names}
\usage{
make_dis_spp(DF, SP_COL)
}
\arguments{
\item{DF}{The dataframe to run the function on}

\item{SP_COL}{The column with the species names}
}
\value{
Returns a dataframe with distinct species names
}
\description{
Gets a list of distinct species names form the
database, and renames the species column
}
\examples{
# example code
 taxon <- make_dis_spp(fm_fish_abundance_clean, Species)

}
