% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_rmv_dups}
\alias{make_rmv_dups}
\title{Take out duplicates for loactions, and taxons}
\usage{
make_rmv_dups(DF, SQL_TBL)
}
\arguments{
\item{DF}{The dataframe to run the function on}

\item{SQL_TBL}{The sql table that matches the dataframe}
}
\value{
Returns the dataframe only records that are not currently in the sql table.
}
\description{
Take out duplicates for loactions, and taxons
}
\examples{
## example code
# taxon_new <- make_rmv_dups(taxon, "dim_taxon")

}
