% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connect.R
\name{sql_con}
\alias{sql_con}
\title{connect to a Postgresql database}
\usage{
sql_con(DBNAME)
}
\arguments{
\item{DBNAME}{is the name of the database you wish to connect to.
It should be a charater with inverted commas.}
}
\value{
a connection object to the sql database provided.
}
\description{
This function will connect to a postgresql database
given a database name. It requires that you have saved the
database port, username, and password in your .Renviron file.
This function works for both windows and linux systems.
}
\details{
To save variables in your .Renviron file open a text file, add
DB_PORT, DB_USER, and DB_PASSWORD with an = after each and the
relevant value. Then run the code, save the file as .Renviron,
and exit and restart R.
}
\examples{
## example code
# con <- sql_con("occur_stage")
}
