% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connect.R
\name{sql_tbl_ext}
\alias{sql_tbl_ext}
\title{extract all tables}
\usage{
sql_tbl_ext(DBNAME, STR)
}
\arguments{
\item{DBNAME}{is the name of the database you wish to connect to.
It should be a charater with inverted commas.}

\item{STR}{is the character string that should occur in all tables that
you wish to retrieve from the database.}
}
\value{
All tables with the matching string as dataframes with the same
names as those in the database. It should be a character in inverted
commas
}
\description{
This function extracts all tables in a postgres database that
match the string provided
}
