% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_area}
\alias{clean_area}
\title{Restrict locations to with the eez}
\usage{
clean_area(DF, LON, LAT, AREA, BUFFER)
}
\arguments{
\item{DF}{The dataframe to run the function on}

\item{LON}{The column with the longitude values (as decimals). Should be a
character in inverted commas.}

\item{LAT}{The column with the latitude values (as decimals). Should be a
character in inverted commas.}

\item{AREA}{The shape file you want to crop your data to. Should be the name
of the shape file plus the file extension.}

\item{BUFFER}{A numeric variable of the buffer zone in meters to be used}
}
\value{
Returns the dataframe with only the values that fall with the eez
}
\description{
This function test if locations fall within an area, with an optional buffer
and drops all other values.
}
\examples{
# example code
 clean_area(species_locations, "longitude", "Latitude", "EEZ_South_Africa_buffered_beyond_allEEZversions1.shp", 2000)
}
