% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_accs}
\alias{extract_accs}
\title{Load tables from access}
\usage{
extract_accs(MDBFILE)
}
\arguments{
\item{MDBFILE}{File name of the access database with the extention.}
}
\value{
Returns all tables in the access database as r dataframes
with the same names.
}
\description{
This function loads tables in an access database into r as
dataframes with the same name as the access tables. If there are multiple
files with the same name it will take the first instance found, which
will be the file closest to the data root folder.
}
\examples{
## example code - this won't run but is for you to see the correct formatting
# extract_accs("FishMaster.accdb")



}
